Low severity - 2 - Audit Report Bug Fix - PantherPoolV0: scaledAmount is already uint96 Since scaledAmount is already declared as uint96, explicit conversion is unnecessary.
Changes are as follows -
- Removed explicit conversion of scaledAmount to uint96.

+++ panther-protocol-contracts-ws/contracts/PantherPoolV0.sol
@@ -171,7 +171,7 @@ contract PantherPoolV0 is
                 commitments[utxoIndex] = generateCommitment(
                     pubSpendingKeys[utxoIndex].x,
                     pubSpendingKeys[utxoIndex].y,
-                    uint96(scaledAmount),
+                    scaledAmount,
                     zAssetId,
                     timestamp
                 );
